interaction point now controlled by transmutation
