+++ Assets/Players/Player.cs
@@ -8,19 +8,16 @@ public class Player : MonoBehaviour
     [Header("Information")]
     [SerializeField] private string propName;
 
-    // These two should be guaranteed to be normalized
+    // This should be guaranteed to be normalized
     private Vector2 moveDirection;
-    private Vector2 mostRecentMoveDirection;
 
     // Components
-    private Interactor interactor;
     private Camera mainCamera;
     private PlayerInputReceiver playerInputReceiver;
-    [SerializeField] private TransmutationBase currentTransmutation;
+    [SerializeField] public TransmutationBase currentTransmutation;
 
     private void Awake()
     {
-        interactor = GetComponent<Interactor>();
         mainCamera = Camera.main;
         playerInputReceiver = GetComponent<PlayerInputReceiver>();
         if (currentTransmutation == null)
@@ -35,7 +32,6 @@ public class Player : MonoBehaviour
     void Update()
     {
         GetMovementInput();
-        UpdateInteractorPosition();
     }
 
     private void FixedUpdate()
@@ -47,15 +43,6 @@ public class Player : MonoBehaviour
     {
         moveDirection = playerInputReceiver.MoveInput;
         moveDirection.Normalize();
-        if (moveDirection.magnitude >= 0.1)
-        {
-            mostRecentMoveDirection = moveDirection;
-        }
-    }
-    
-    void UpdateInteractorPosition()
-    {
-        interactor.interactionPoint.localPosition = mostRecentMoveDirection;
     }
 
     public bool Transmute(TransmutationBase transmutation)
