--- Assets/Props/IInteractable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 48f195d17468e594bbf66dc1fc02a1af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
