--- Assets/Props/Prop.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 89d8204044b9ea54cab4fb5b27db74c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
