+++ Assets/Props/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a80e68521e899f4409b912dd21209874
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
